refactor: Simplify collector configuration and content processing
- Extract content processing logic into separate function
- Add parameter validation function for collector initialization
- Remove redundant code and improve readability
- Maintain existing collector functionality with cleaner implementation
