+++ internal/collector/collector.go
@@ -1,97 +1,17 @@
 package collector
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	"net/url"
 	"strings"
-	"time"
-
-	"github.com/jonesrussell/gocrawl/internal/article"
-	"github.com/jonesrussell/gocrawl/internal/logger"
-	"github.com/jonesrussell/gocrawl/internal/sources"
 
 	"github.com/gocolly/colly/v2"
-	"github.com/gocolly/colly/v2/debug"
-	"go.uber.org/fx"
 )
-
-// DebuggerInterface is an interface for the debugger
-type DebuggerInterface interface {
-	Init() error
-	OnRequest(e *colly.Request)
-	OnResponse(e *colly.Response)
-	OnError(e *colly.Response, err error)
-	OnEvent(e *debug.Event)
-	Event(e *debug.Event)
-}
-
-// ContentProcessor processes non-article content
-type ContentProcessor interface {
-	ProcessContent(e *colly.HTMLElement)
-}
-
-const (
-	// Default selectors when none are specified in the source config
-	DefaultArticleSelector    = "body"
-	DefaultTitleSelector      = "h1"
-	DefaultDateSelector       = "time"
-	DefaultAuthorSelector     = "span.author"
-	DefaultCategoriesSelector = "div.categories"
-
-	// Context keys
-	articleFoundKey = "articleFound"
-	bodyElementKey  = "bodyElement"
-)
-
-// getSelector returns the specified selector or falls back to a default
-func getSelector(specified, defaultSelector string) string {
-	if specified == "" {
-		return defaultSelector
-	}
-	return specified
-}
-
-// Params holds the parameters for creating a Collector
-type Params struct {
-	fx.In
-
-	ArticleProcessor *article.Processor
-	ContentProcessor ContentProcessor
-	BaseURL          string
-	Context          context.Context
-	Debugger         *logger.CollyDebugger
-	Logger           logger.Interface
-	MaxDepth         int
-	Parallelism      int
-	RandomDelay      time.Duration
-	RateLimit        time.Duration
-	Source           *sources.Config
-}
-
-// Result holds the collector instance
-type Result struct {
-	fx.Out
-
-	Collector *colly.Collector
-}
 
 // New creates a new collector instance
 func New(p Params) (Result, error) {
-	// Validate URL
-	if p.BaseURL == "" {
-		return Result{}, errors.New("base URL cannot be empty")
-	}
-
-	// Check if ArticleProcessor is nil
-	if p.ArticleProcessor == nil {
-		return Result{}, errors.New("article processor is required")
-	}
-
-	// Check if Logger is nil
-	if p.Logger == nil {
-		return Result{}, errors.New("logger is required")
+	if err := ValidateParams(p); err != nil {
+		return Result{}, err
 	}
 
 	parsedURL, err := url.Parse(p.BaseURL)
@@ -106,7 +26,7 @@ func New(p Params) (Result, error) {
 	c := colly.NewCollector(
 		colly.Async(true),
 		colly.MaxDepth(p.MaxDepth),
-		colly.AllowedDomains(domain), // Set the allowed domain
+		colly.AllowedDomains(domain),
 	)
 
 	// Set rate limiting
@@ -126,51 +46,8 @@ func New(p Params) (Result, error) {
 	// Configure logging
 	ConfigureLogging(c, p.Logger)
 
-	// Set up link following
-	var ignoredErrors = map[string]bool{
-		"Max depth limit reached": true,
-		"Forbidden domain":        true,
-		"URL already visited":     true,
-	}
-
-	c.OnHTML("a[href]", func(e *colly.HTMLElement) {
-		link := e.Attr("href")
-		p.Logger.Debug("Link found", "text", e.Text, "link", link)
-		visitErr := e.Request.Visit(e.Request.AbsoluteURL(link))
-		if visitErr != nil && !ignoredErrors[visitErr.Error()] {
-			p.Logger.Error("Failed to visit link", "link", link, "error", visitErr)
-		}
-	})
-
-	// Get the article selector with fallback
-	articleSelector := getSelector(p.Source.Selectors.Article, DefaultArticleSelector)
-
-	// Store body element when found
-	c.OnHTML("body", func(e *colly.HTMLElement) {
-		e.Request.Ctx.Put(bodyElementKey, e)
-	})
-
-	c.OnHTML(articleSelector, func(e *colly.HTMLElement) {
-		e.Request.Ctx.Put(articleFoundKey, "true")
-		p.Logger.Debug("Found article", "url", e.Request.URL.String(), "selector", articleSelector)
-		p.ArticleProcessor.ProcessPage(e)
-	})
-
-	// Handle non-article content after response is received
-	c.OnResponse(func(r *colly.Response) {
-		if r.Ctx.Get(articleFoundKey) != "true" && p.ContentProcessor != nil {
-			p.Logger.Debug("Processing non-article content", "url", r.Request.URL.String())
-			// Create a temporary collector to parse the response
-			temp := colly.NewCollector()
-			temp.OnHTML("body", func(e *colly.HTMLElement) {
-				p.ContentProcessor.ProcessContent(e)
-			})
-			err := temp.Visit("data:text/html;charset=utf-8," + string(r.Body))
-			if err != nil {
-				p.Logger.Error("Failed to process non-article content", "url", r.Request.URL.String(), "error", err)
-			}
-		}
-	})
+	// Configure content processing
+	configureContentProcessing(c, p)
 
 	p.Logger.Debug("Collector created",
 		"baseURL", p.BaseURL,
@@ -180,25 +57,4 @@ func New(p Params) (Result, error) {
 	)
 
 	return Result{Collector: c}, nil
-}
-
-// ConfigureLogging sets up logging for the collector
-func ConfigureLogging(c *colly.Collector, log logger.Interface) {
-	c.OnRequest(func(r *colly.Request) {
-		log.Debug("Requesting URL", "url", r.URL.String())
-	})
-
-	c.OnResponse(func(r *colly.Response) {
-		log.Debug("Received response",
-			"url", r.Request.URL.String(),
-			"status", r.StatusCode,
-		)
-	})
-
-	c.OnError(func(r *colly.Response, err error) {
-		log.Error("Error occurred",
-			"url", r.Request.URL.String(),
-			"error", err.Error(),
-		)
-	})
 }
